requirements.txt
selenium==4.3.0
pytest==7.1.2
webdriver-manager==3.8.2
python-dotenv==0.20.0